// ts-source/src/enums/role.ts
export enum Role {
    Protector = 0,
    RampartDefender = 1,
    BaseRanger = 2,
    Ranger = 3,
    OutpostSupporter = 4,
    OutpostWarrior = 5,
    BaseHauler = 6,
    ProximityScout = 7,
    OutpostReserver = 8,
    Pioneer = 9,
    RoomClaimer = 10,
    MineralMiner = 11,
    ExpansionWorker = 12,
    ContainerMiner = 13,
    ContainerHauler = 14,
    OutpostDefender = 15,
    Janitor = 16,
    Upgrader = 17,
    PoachGuard = 18,
    TeamHealer = 19,
    TeamWarrior = 20,
    TeamWrecker = 21,
    PoachMiner = 22,
    PoachHauler = 23,
    BaseBuilder = 24,
    Harasser = 25,
    PraiserSupport = 26,
    SKGuard = 27,
    BankAttacker = 28,
    BankHealer = 29,
    BankHauler = 30,
    BaseCourier = 31,
    SKMiner = 32,
    SKHauler = 33,
    SKHealer = 34,
    BankRanger = 35,
    MineralHauler = 36,
    EnergyHauler = 37,
    Signer = 38,
    PraiserOfficer = 39,
    /* UNUSED = 40,*/
    Wrecker = 41,
    Drainer = 42,
    Paladin = 43,
    Praiser = 44,
    PraiserHauler = 45,
    Tagger = 46,
    PraiserLeader = 47,
    Declarer = 48,
    UpgraderHauler = 49,
    OperationHauler = 50,

    UpgraderWithBoost = 1001,
    UpgraderWithoutBoost = 1002,
    PraiserWithBoost = 1003,
    PraiserWithoutBoost = 1004
}